typed options && move chartTypes in it's own file
